+++ zebrad/src/components/mempool.rs
@@ -40,8 +40,6 @@ use self::downloads::{
     Downloads as TxDownloads, Gossip, TRANSACTION_DOWNLOAD_TIMEOUT, TRANSACTION_VERIFY_TIMEOUT,
 };
 
-#[cfg(test)]
-use super::sync::RecentSyncLengths;
 use super::sync::SyncStatus;
 
 type Outbound = Buffer<BoxService<zn::Request, zn::Response, zn::BoxError>, zn::Request>;
@@ -173,44 +171,6 @@ impl Mempool {
             ActiveState::Disabled => false,
             ActiveState::Enabled { .. } => true,
         }
-    }
-
-    /// Get the storage field of the mempool for testing purposes.
-    #[cfg(test)]
-    pub fn storage(&mut self) -> &mut storage::Storage {
-        match &mut self.active_state {
-            ActiveState::Disabled => panic!("mempool must be enabled"),
-            ActiveState::Enabled { storage, .. } => storage,
-        }
-    }
-
-    /// Get the transaction downloader of the mempool for testing purposes.
-    #[cfg(test)]
-    pub fn tx_downloads(&self) -> &Pin<Box<InboundTxDownloads>> {
-        match &self.active_state {
-            ActiveState::Disabled => panic!("mempool must be enabled"),
-            ActiveState::Enabled { tx_downloads, .. } => tx_downloads,
-        }
-    }
-
-    /// Enable the mempool by pretending the synchronization is close to the tip.
-    #[cfg(test)]
-    pub async fn enable(&mut self, recent_syncs: &mut RecentSyncLengths) {
-        use tower::ServiceExt;
-        // Pretend we're close to tip
-        SyncStatus::sync_close_to_tip(recent_syncs);
-        // Make a dummy request to poll the mempool and make it enable itself
-        let _ = self.oneshot(Request::TransactionIds).await;
-    }
-
-    /// Disable the mempool by pretending the synchronization is far from the tip.
-    #[cfg(test)]
-    pub async fn disable(&mut self, recent_syncs: &mut RecentSyncLengths) {
-        use tower::ServiceExt;
-        // Pretend we're far from the tip
-        SyncStatus::sync_far_from_tip(recent_syncs);
-        // Make a dummy request to poll the mempool and make it disable itself
-        let _ = self.oneshot(Request::TransactionIds).await;
     }
 
     /// Check if transaction should be downloaded and/or verified.
